+++ CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.34.0](https://github.com/nearform/orion/compare/v2.33.0...v2.34.0) (2020-03-24)
+
+
+### Features
+
+* 🎸 drag and drop management of page hierarchy ([0d520da](https://github.com/nearform/orion/commit/0d520da7c9820c4e08e36393df0da953c27ffcee))
+
+
+
+
+
+# [2.33.0](https://github.com/nearform/orion/compare/v2.32.1...v2.33.0) (2020-03-24)
+
+
+### Features
+
+* 🎸 drag and drop tree ui for page management ([8953efe](https://github.com/nearform/orion/commit/8953efe6f0356f86ce716ac57c4f43a0dc41e96b))
+* 🎸 re-ordering of pages within the existing hierarchy ([652fb4e](https://github.com/nearform/orion/commit/652fb4e90fdd6c7352adba16d1e3a3d52e760aaf))
+
+
+
+
+
+## [2.32.1](https://github.com/nearform/orion/compare/v2.32.0...v2.32.1) (2020-03-24)
+
+
+### Bug Fixes
+
+* 🐛 Allow for content of an empty string ([8965a57](https://github.com/nearform/orion/commit/8965a576256697fbf166ca07423f05b1b8d7de09))
+* 🐛 Typo in propTypes and accidental removal of react-helmet ([6ec46ca](https://github.com/nearform/orion/commit/6ec46caa76fe6ff93d3c423d76b38d6478bd4d8d))
+
+
+### Features
+
+* 🎸 Add title to pages ([67b6293](https://github.com/nearform/orion/commit/67b62931389806f02fad9dc6151cc3a4d59b6694))
+
+
+
+
+
 # [2.32.0](https://github.com/nearform/orion/compare/v2.31.0...v2.32.0) (2020-03-19)